[dr] use coroutines

+++ sources/dependency-resolution/src/org/jetbrains/amper/dependency/resolution/metadata/json/json.kt
@@ -101,7 +101,7 @@ data class Version(
 data class File(
     val name: String,
     val url: String,
-    val size: Int? = null,
+    val size: Long? = null,
     val sha512: String? = null,
     val sha256: String? = null,
     val sha1: String? = null,
